kubernetes: prevent port conflicts with apiserver
Pick a random free port on the host and use that as the value for
the `--https-listen-port` argument for k3s.

This makes it possible to create multiple colima profiles/instances
with Kubernetes; otherwise, they'll all be listening on 6443, and
only the first one will be port forwarded, so attempting to use any
others will result in x509 errors since the config<>cluster mapping
is erroneous.