docker: Increase ulimit to infinity
The current Colima VM (Ubuntu) is using the default value of 1024 file descriptors. This is insufficient for software such as Kafka that requires a high amount of open files, even in development environments.

https://docs.confluent.io/platform/current/kafka/deployment.html

Since the principal purpose of the Colima VM is just to run docker, this sets the value to infinity to limit to the OS limits.