Write kubeconfig to KUBECONFIG defined file
Currently when starting a vm + k8s cluster the kubeconfig is always written to "~/.kube/config" which is the default path.
By defining the env variable "KUBECONFIG" it is possible to override the default path.
This does currently not apply to colima.
As it is possible to configure multiple locations with "KUBECONFIG" I opted to always use the first configured one as the location to save the config.