fix(docker): apply configuration on first boot
The initialization sequence is:
 (1) VM boot
 (2) Write daemon configuration

When the VM boot, the Docker systemd service is started, so the
new configuration is not picked up until the next VM boot.

To fix this, reload the systemd daemon config (to pick up the
systemd service file written by `addHostGateway`) and then restart
the Docker systemd service, which will use the new systemd unit
file as well as any customizations to `daemon.json`.