vm: make binfmt emulation configurable (#1315)
As ed49fef4ad63ec48055b4b230ecbad00c54f5417 did for Rosetta, this change
adds a flag to control whether binfmt emulation is setup.

In the future, I think it'd be interesting to consider making the
option's type `[]string | bool`, where `true` keeps the default behavior
of automatically picking the architectures to enable, `false` disables
binfmt setup, and a list will let the user define which architectures to
emulate.